Moved backend to use TRPC. The Elysia integration remains present until we are sure all functionality works. Right now, TRPC works for resource fetch confirmed.
The reason we chose TRPC is due to its use of zod vs. a custom typebox
implementation in Elysia, as well as DX with its ability to cmd+click
into the right procedure. The new Tanstack Query integration also helps
and we can dump the /api folder from our feature architecture with
TRPC.